% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_bh_adjust}
\alias{fast_bh_adjust}
\title{Fast P-value Adjustment (C++)}
\usage{
fast_bh_adjust(p_values)
}
\arguments{
\item{p_values}{Vector of p-values}
}
\value{
Vector of adjusted p-values
}
\description{
Fast implementation of Benjamini-Hochberg FDR correction.
}
