% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_module_score_cpp}
\alias{fast_module_score_cpp}
\title{Fast Module Score Calculation (C++)}
\usage{
fast_module_score_cpp(
  cor_matrix,
  variation,
  module_genes,
  all_genes,
  add_size = FALSE
)
}
\arguments{
\item{cor_matrix}{Correlation matrix}

\item{variation}{Vector of gene variations}

\item{module_genes}{Indices of genes in module (0-based)}

\item{all_genes}{Total number of genes}

\item{add_size}{Whether to include module size in score}
}
\value{
List with CI, V_in, R_in, R_out
}
\description{
Optimized calculation of DNB module scores.
}
