% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_sspn_batch}
\alias{fast_sspn_batch}
\title{Fast SSPN Calculation for All Samples (C++)}
\usage{
fast_sspn_batch(
  expr_mat,
  ref_indices,
  case_indices,
  ppi_gene1,
  ppi_gene2,
  gene_names
)
}
\arguments{
\item{expr_mat}{Expression matrix (genes x all_samples)}

\item{ref_indices}{Indices of reference samples (0-based)}

\item{case_indices}{Indices of case samples (0-based)}

\item{ppi_gene1}{First gene names in PPI}

\item{ppi_gene2}{Second gene names in PPI}

\item{gene_names}{Gene names in expression matrix}
}
\value{
List of perturbation data frames for each case sample
}
\description{
Batch calculation of SSPN for all case samples.
Major speedup by avoiding repeated matrix operations.
}
\keyword{internal}
