% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDNB.R
\name{MDNB}
\alias{MDNB}
\title{Module-based Dynamic Network Biomarker (MDNB) Analysis}
\usage{
MDNB(
  expr,
  state,
  state.levels,
  cor.method = "pearson",
  ppi = ppi_h,
  min.combined.score = 900,
  PCC.min = 0.02,
  PCC.module = 0.02,
  numeber.module.QI = 10
)
}
\arguments{
\item{expr}{A expression dataframe with gene rows and sample columns.}

\item{state}{A dataframe with two columns: sample names and group information.}

\item{state.levels}{A vector for state sequence (e.g., cell groups, time points, treatment conditions).}

\item{cor.method}{Specifies the method for correlation analysis.}

\item{ppi}{Protein-protein interaction network; background network.}

\item{min.combined.score}{Minimum combined score for determining protein-protein interaction.}

\item{PCC.min}{Remove genes if their highest correlation with any other gene is less than this value.}

\item{PCC.module}{A gene must maintain at least this minimum correlation with the seed gene
in every group to be included in the module.}

\item{numeber.module.QI}{Number of gene modules selected for QI calculation.}
}
\value{
A list containing:
\itemize{
\item DNB.score: The DNB score values (QI values) calculated for different groups.
\item DNB.genes: Genes from the module with highest CI in the critical state.
\item CI_all: Module contribution index (CI) details for each state.
\item Gene_module: Generated co-expression gene modules.
\item QIplot: Line chart showing QI changes across states.
\item metadata: Additional information about the analysis.
}
}
\description{
Performing module-based dynamic network biomarker analysis using PPI network
}
\references{
Li L, Xu Y, Yan L, et al. Dynamic network biomarker factors orchestrate cell-fate
determination at tipping points during hESC differentiation. Innovation (Camb).
2022 Dec 20;4(1):100364. doi: 10.1016/j.xinn.2022.100364
}
\author{
Zaoqu Liu, Chuhan Zhang; Email: liuzaoqu@163.com
}
