% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNMB.R
\name{sNMB}
\alias{sNMB}
\title{Single-sample network module biomarker analysis}
\usage{
sNMB(
  expr,
  state,
  state.levels,
  cor.method = "pearson",
  p.adjust.method = "BH",
  ppi = ppi_h,
  min.combined.score = 900,
  min.first.neighbor.size = 3,
  percent = TRUE,
  top.n = 30,
  top.p = 0.05,
  nCores = parallel::detectCores() - 10
)
}
\arguments{
\item{expr}{A expression dataframe with gene rows and sample columns.}

\item{state}{A time-series dataframe with two columns, the first is the sample names and the second is the group or time point information.}

\item{state.levels}{A vector for state sequence.}

\item{cor.method}{specifies the method for correlation analysis.}

\item{p.adjust.method}{correction method, a character string. Can be abbreviated. c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}

\item{ppi}{Protein-protein interaction network; background network.}

\item{min.combined.score}{Minimum combined score for determining protein-protein interaction.}

\item{min.first.neighbor.size}{Minimum size of first order genes of a specific center gene.}

\item{percent}{Whether to use Percent to determine the number of DNB genes.}

\item{top.n}{Only percent = FALSE takes effect. Center genes with top (number) DNB score were defined as DNB genes.}

\item{top.p}{Only percent = TRUE takes effect. Center genes with top (percent) DNB score were defined as DNB genes.}

\item{nCores}{The number of cores will be used.}
}
\description{
Performing single-sample network module biomarker analysis.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
