% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSPN.R
\name{SSPN2}
\alias{SSPN2}
\title{Sample-specific perturbation network based on customized network}
\usage{
SSPN2(
  expr,
  ref.samples,
  net,
  cor.method = "pearson",
  p.adjust.method = "BH",
  nCores = max(1L, parallel::detectCores() - 1L)
)
}
\arguments{
\item{expr}{A expression dataframe with gene rows and sample columns.}

\item{ref.samples}{Samples for constructing the background reference network; Column ID or names.}

\item{net}{A customized network with two columns.}

\item{cor.method}{specifies the method for correlation analysis.}

\item{p.adjust.method}{correction method, a character string. Can be abbreviated. c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}

\item{nCores}{The number of cores will be used.}
}
\description{
Performing sample-specific perturbation network analysis based on customized network.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
