% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDNB.R
\name{LDNB}
\alias{LDNB}
\title{Landscape DNB analysis}
\usage{
LDNB(
  expr,
  state,
  state.levels,
  cor.method = "pearson",
  p.adjust.method = "BH",
  ppi = ppi_h,
  min.combined.score = 990,
  min.first.neighbor.size = 10,
  min.second.neighbor.size = 1,
  use.PCC.P.type = "FDR",
  use.PCC.P.cutoff = 0.05,
  percent = F,
  top.n = 30,
  top.p = 0.05,
  nCores = parallel::detectCores() - 10
)
}
\arguments{
\item{expr}{A expression dataframe with gene rows and sample columns.}

\item{state}{A time-series dataframe with two columns, the first is the sample names and the second is the group or time point information.}

\item{state.levels}{A vector for state sequence.}

\item{cor.method}{specifies the method for correlation analysis.}

\item{p.adjust.method}{correction method, a character string. Can be abbreviated. c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}

\item{ppi}{Protein-protein interaction network; background network.}

\item{min.combined.score}{Minimum combined score for determining protein-protein interaction.}

\item{min.first.neighbor.size}{Minimum size of first order genes of a specific center gene.}

\item{min.second.neighbor.size}{Minimum size of second order genes of a specific center gene.}

\item{use.PCC.P.type}{Nominal P value (NP) or adjust P value (FDR) were selected to define the significance of sPCC.}

\item{use.PCC.P.cutoff}{A cutoff value for use.PCC.P.type.}

\item{percent}{Whether to use Percent to determine the number of DNB genes.}

\item{top.n}{Only Percent = F takes effect. Center genes with top (number) DNB score were defined as DNB genes.}

\item{top.p}{Only Percent = T takes effect. Center genes with top (percent) DNB score were defined as DNB genes.}

\item{nCores}{The number of cores will be used.}

\item{variation.method}{specifies the method for calculating gene variation. sd or cv.}
}
\description{
Performing landscape dynamic network biomarker analysis.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
